feat(maszyny): podepnij pomieszczenia bez przebudowy modułu
